services/requirements.txt
fastapi
uvicorn
pydantic
google-generativeai
python-multipart
sqlalchemy
redis
aiofiles
